Moving the rest of the browser detection functions over.
